+++ iowmongotools/__init__.py
@@ -12,6 +12,9 @@ logger = logging.getLogger(__name__)
 class MongoCheckerCli(app.AppCli):
 
     def run(self):
+        if not hasattr(self.config, 'clusters'):
+            logger.error('Please provide cluster_config.yaml. See --help.')
+            return 1
         errors = len(self.config.clusters) - cluster.create_objects(self.config.clusters, self.config.cluster_config)
         for name, _cluster in cluster.Cluster.objects.items():
             if not _cluster.check_config():