+++ tests/conftest.py
@@ -1,16 +1,16 @@
 # content of conftest.py
+import os
 import pytest
 from iowmongotools import cluster
-
+import yaml
 
 @pytest.fixture(scope="module")
-# Start mongo in docker. docker run -d --rm -p 27017:27017 mongo
+# Use pymongo for real mongo. Start mongo in docker: docker run -d --rm -p 27017:27017 mongo
 def local_cluster():
-    import yaml
-    import pymongo
-    client = pymongo.MongoClient()
+    import mongomock
+    client = mongomock.MongoClient()
     # seed sample db config
-    with open('sample_cluster.yaml', 'r') as f:
+    with open(os.path.join(os.path.dirname(__file__), 'sample_cluster.yaml'), 'r') as f:
         cluster_config = yaml.safe_load(f.read())['gce-eu']
     for key in cluster_config:
         client.config[key].drop()